:truck: Move flow models to the `lib/models`.
