+++ frontend/src/components/Paths/index.tsx
@@ -2,7 +2,7 @@ import { observer } from 'mobx-react-lite';
 
 import { Path } from '@/components/Path';
 import { useDiagramNodes } from '@/diagrams';
-import { Path as PathType } from '@/lib/Path';
+import { Path as PathType } from '@/lib/models/Path';
 
 export const Paths: React.VFC<{ paths: PathType[] }> = observer(({ paths }) => {
   const nodes = useDiagramNodes();
