+++ frontend/src/components/Task/index.tsx
@@ -2,7 +2,7 @@ import { observer } from 'mobx-react-lite';
 import { useRef } from 'react';
 
 import { useRegisterNode } from '@/diagrams';
-import { Task as TaskType } from '@/lib/Task';
+import { Task as TaskType } from '@/lib/models/Task';
 
 export const Task: React.VFC<{ task: TaskType }> = observer(({ task }) => {
   const taskRef = useRef<SVGRectElement>(null);
