+++ frontend/src/lib/models/Aggregate.ts
