+++ frontend/src/lib/models/Path.ts
